+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog for Windows-Installer-NPX
 <!-- **WORK IN PROGRESS** -->
+**WORK IN PROGRESS**
+* Allowed to install on linux too
+
 ## 4.0.3 (2022-05-22)
 * Corrected fixer